requirements.txt
Django==1.10.4
psycopg2==2.8.6
djangorestframework==3.9.4